+++ scm-ui/ui-types/src/Me.ts
@@ -27,7 +27,8 @@ import { Links } from "./hal";
 export type Me = {
   name: string;
   displayName: string;
-  mail: string;
+  mail?: string;
+  fallbackMail?: string;
   groups: string[];
   _links: Links;
 };