+++ CHANGELOG.md
@@ -7,7 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- Generation of email addresses for users, where none is configured ([#1370](https://github.com/scm-manager/scm-manager/pull/1370))
 - Source code fullscreen view ([#1376](https://github.com/scm-manager/scm-manager/pull/1376))
+
+### Fixed
+- Missing default permission to manage public gpg keys ([#1377](https://github.com/scm-manager/scm-manager/pull/1377))
+
+## [2.6.3] - 2020-10-16
+### Fixed
+- Missing default permission to manage public gpg keys ([#1377](https://github.com/scm-manager/scm-manager/pull/1377))
 
 ## [2.7.1] - 2020-10-14
 ### Fixed
@@ -362,3 +370,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [2.6.0]: https://www.scm-manager.org/download/2.6.0
 [2.6.1]: https://www.scm-manager.org/download/2.6.1
 [2.6.2]: https://www.scm-manager.org/download/2.6.2
+[2.6.3]: https://www.scm-manager.org/download/2.6.3
+[2.7.0]: https://www.scm-manager.org/download/2.7.0
+[2.7.1]: https://www.scm-manager.org/download/2.7.1